A2A协议深度解析(2)/flight/pyproject.toml
[project]
name = "flight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "uvicorn>=0.34.2",
]
